+++ src/TypeShape/Utils.fs
@@ -5,51 +5,67 @@ module internal TypeShape_Utils
 #endif
 
 open System
-open System.Collections.Generic
+open System.Collections.Concurrent
 
-type private Cell<'T>() =
-    let mutable value = Unchecked.defaultof<'T>
-    let mutable isValueSet = false
-    member __.IsValueSet = isValueSet
-    member __.Value =
-        if isValueSet then value
-        else failwithf "Value for %O has not been initialized." typeof<'T>
+/// Value container that will eventually be populated
+type Cell<'T> internal (container : 'T option ref) =
+    member __.IsValueCreated = Option.isSome !container
+    member __.Value = 
+        match container.Value with
+        | None -> failwithf "Value for '%O' has not been initialized." typeof<'T>
+        | Some t -> t
 
-    member __.SetValue(t:'T) = value <- t ; isValueSet <- true
+type private Payload<'T> = 'T option ref * 'T
 
-type private Payload<'T> = Cell<'T> * 'T
+/// Thread-safe cache of values indexed by type that supports recursion.
+type TypedIndex() =
+    let dict = new ConcurrentDictionary<Type, obj>()
 
-/// State object for managing recursive types
-type RecTypeManager() =
-    let dict = new Dictionary<Type, obj>()
+    /// <summary>
+    ///     Registers an uninitialized value at the beggining of a recursive
+    ///     value definition. Returns a dummy value that can be referenced within a
+    ///     recursive flow. Only delayable values can be recursive.
+    /// </summary>
+    /// <param name="delay">Provides delay wrapping for supplied type.</param>
+    member __.CreateUninitialized<'T>(delay : Cell<'T> -> 'T) : 'T =
+        let create _ =
+            let container = ref None
+            let dummy = delay (Cell container)
+            let payload : Payload<'T> = container, dummy
+            box payload
 
-    /// Try to get value for spefic type
-    member __.TryGetValue<'T>() =
-        let ok, obj = dict.TryGetValue typeof<'T>
-        if ok then
-            let container, delayed = obj :?> Payload<'T>
-            if container.IsValueSet then Some container.Value
-            else Some delayed
+        let _, value = dict.GetOrAdd(typeof<'T>, create) :?> Payload<'T>
+        value
+
+    /// Registers a value to the type index. Any uninitialized references 
+    /// to this type will be updated to point to this value.
+    member __.Commit<'T>(value : 'T) : 'T =
+        let create _ =
+            let payload : Payload<'T> = ref (Some value), value
+            box payload
+
+        let update _ (currentValue : obj) =
+            let container, _ = currentValue :?> Payload<'T>
+            lock container (fun () ->
+                match container.Value with
+                | Some _ -> currentValue
+                | None ->
+                    container := Some value
+                    let newPayload : Payload<'T> = container, value
+                    box newPayload)
+
+        let _,value = dict.AddOrUpdate(typeof<'T>, create, update) :?> Payload<'T>
+        value
+
+    member __.TryGetValue<'T>(result : byref<'T>) : bool =
+        let mutable obj = null
+        if dict.TryGetValue(typeof<'T>, &obj) then
+            let _,value = obj :?> Payload<'T>
+            result <- value ; true
         else
-            None
+            false
 
-    /// Creates a dummy value for given type using supplied delay function
-    member __.Create<'T>(delay : (unit -> 'T) -> 'T) : 'T =
-        if dict.ContainsKey typeof<'T> then
-            failwithf "A definition for '%O' has already been created." typeof<'T>
-        let container = new Cell<'T>()
-        let delayed = delay (fun () -> container.Value)
-        let payload : Payload<'T> = container, delayed
-        dict.[typeof<'T>] <- box payload
-        delayed
-
-    /// Completes recursive definition for given type using supplied value
-    member __.Complete<'T>(value : 'T) =
-        let ok, obj = dict.TryGetValue typeof<'T>
-        if ok then
-            let container, _ = obj :?> Payload<'T>
-            if container.IsValueSet then
-                failwithf "A value for '%O' has already been set." typeof<'T>
-            container.SetValue value
-        else
-            failwithf "No definition for '%O' has been created." typeof<'T>
+    member __.TryFind<'T>() =
+        let mutable t = Unchecked.defaultof<_>
+        if __.TryGetValue<'T>(&t) then Some t
+        else None